src/requirements.txt
kubernetes==29.0.0
google-cloud-compute==1.18.0